% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{base_plot}
\alias{base_plot}
\title{Wrapper to a default ggplot for voronoise}
\usage{
base_plot(data = base_data(), background = "antiquewhite3")
}
\arguments{
\item{data}{data for voronoi tesselation}

\item{background}{background colour of the plot}
}
\value{
a ggplot object
}
\description{
Wrapper to a default ggplot for voronoise
}
