% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{coord_square}
\alias{coord_square}
\title{Like coord_fixed() but forced to be a square}
\usage{
coord_square(x = 0, y = 0, s = 1, ...)
}
\arguments{
\item{x}{bottom side of the square}

\item{y}{left side of the square}

\item{s}{side length}

\item{...}{arguments to be passed to coord_fixed()}
}
\description{
Like coord_fixed() but forced to be a square
}
