% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_voronoise}
\alias{geom_voronoise}
\title{Voronoise geom}
\usage{
geom_voronoise(
  mapping = NULL,
  data = NULL,
  stat = "voronoise_tile",
  perturb = identity,
  ...,
  show.legend = FALSE,
  max.radius = 0.1,
  radius = 0.01,
  expand = -0.005
)
}
\arguments{
\item{mapping}{dfadsf}

\item{data}{dfsdf}

\item{stat}{asdfasdf}

\item{...}{fdssdf}

\item{show.legend}{sdfsdf}

\item{max.radius}{adsfasdf}

\item{radius}{adsfsadf}

\item{expand}{asdfasdf}

\item{peturb}{sdfsdf}
}
\description{
Voronoise geom
}
