% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturb.R
\name{perturb_float}
\alias{perturb_float}
\title{Perturb voronoi tiles by drifting along an axis}
\usage{
perturb_float(angle = 90, bound = TRUE, noise = NULL, ...)
}
\arguments{
\item{angle}{direction of the drift}

\item{bound}{should the drift be bounded at the edge of the box}

\item{noise}{function to generate random number on the unit interval (defaults to beta density)}

\item{...}{arguments to be passed to noise}
}
\value{
A function that takes data as input
}
\description{
Perturb voronoi tiles by drifting along an axis
}
