% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturb.R
\name{perturb_uniform}
\alias{perturb_uniform}
\title{Perturb voronoise tiles uniformly}
\usage{
perturb_uniform(noise = 0.1)
}
\arguments{
\item{noise}{width of the uniform}
}
\value{
A function that takes data as input
}
\description{
Perturb voronoise tiles uniformly
}
