% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peturb.R
\name{peturb_uniform}
\alias{peturb_uniform}
\title{Peturb voronoise tiles uniformly}
\usage{
peturb_uniform(data, noise = 0.1)
}
\arguments{
\item{data}{data passed from the voronoi_tile stat}

\item{noise}{width of the uniform}
}
\value{
peturbed data
}
\description{
Peturb voronoise tiles uniformly
}
