% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stat.R
\docType{data}
\name{StatVoronoiseTile}
\alias{StatVoronoiseTile}
\title{Voronoise stat}
\format{
An object of class \code{StatVoronoiseTile} (inherits from \code{StatVoronoiTile}, \code{Stat}, \code{ggproto}, \code{gg}) of length 2.
}
\usage{
StatVoronoiseTile
}
\description{
The voronoise package provides the voronoise_tile stat.
}
\keyword{datasets}
