% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{theme_mono}
\alias{theme_mono}
\title{Like theme_void() but with custom background colour}
\usage{
theme_mono(background = "black")
}
\arguments{
\item{background}{the background colour}
}
\description{
Like theme_void() but with custom background colour
}
