% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{voronoise_base}
\alias{voronoise_base}
\title{Wrapper to a default ggplot for voronoise}
\usage{
voronoise_base(data = voronoise_data(), background = "antiquewhite3")
}
\arguments{
\item{data}{data for voronoi tesselation}

\item{background}{background colour of the plot}
}
\value{
a ggplot object
}
\description{
Wrapper to a default ggplot for voronoise
}
