% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/base.R
\name{voronoise_data}
\alias{voronoise_data}
\title{Generate random points for the voronoi tesselation}
\usage{
voronoise_data(n = 50)
}
\arguments{
\item{n}{Number of points}
}
\value{
Data frame with columns x, y and shade
}
\description{
Generate random points for the voronoi tesselation
}
