% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geom.R
\name{geom_voronoise}
\alias{geom_voronoise}
\title{Voronoise geom}
\usage{
geom_voronoise(
  mapping = NULL,
  data = NULL,
  stat = "voronoise_tile",
  perturb = identity,
  ...,
  show.legend = FALSE,
  max.radius = 0.1,
  radius = 0.01,
  expand = -0.005
)
}
\arguments{
\item{mapping}{ggplot2 mapping specified by aes()}

\item{data}{data frame containing the values to use}

\item{stat}{the statistic to compute}

\item{perturb}{a perturbing function}

\item{...}{arguments to be passed to geom_voronoi_tile()}

\item{show.legend}{by default no legend is displayed}

\item{max.radius}{specifies the maximum distance from tile to point}

\item{radius}{radius specifying curvature of the tile corners}

\item{expand}{space between adjacent tiles}
}
\description{
Voronoise geom
}
