% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/perturb.R
\name{perturb_float}
\alias{perturb_float}
\title{Perturb voronoi tiles by drifting along an axis}
\usage{
perturb_float(angles = 90, bound = TRUE, noise = c(1, 5))
}
\arguments{
\item{angles}{possible directions for the drift in degrees}

\item{bound}{should the drift be bounded at the edge of the box}

\item{noise}{parameters specifying the noise}
}
\value{
A function that takes data as input
}
\description{
Perturb voronoi tiles by drifting along an axis
}
