% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/peturb.R
\name{perturb_uniform}
\alias{perturb_uniform}
\title{Perturb voronoise tiles uniformly}
\usage{
perturb_uniform(data, noise = 0.1)
}
\arguments{
\item{data}{data passed from the voronoi_tile stat}

\item{noise}{width of the uniform}
}
\value{
perturbed data
}
\description{
Perturb voronoise tiles uniformly
}
